chore/dashboard_cleanup
