+++ src/components/Create/DareChallenge/DareChallengeForm.jsx
@@ -7,6 +7,9 @@ import { yellowarrow } from '../../../assets/images'
 import Popup from '../../Popup/Popup'
 import { createChallengeAPI } from '../../../utils/ApiCalls'
 import { useNavigate, useForm } from 'react-router-dom'
+import { useDispatch } from 'react-redux'
+import { setPopupState } from '../../../redux/actions/actions'
+
 
 const DareChallengeForm = ({
   register,
@@ -22,9 +25,9 @@ const DareChallengeForm = ({
   const [isPopupOpen, setIsPopupOpen] = useState(false)
   const [cid, setCid] = useState('')
   const navigate = useNavigate()
-
-  // Watch the challengeType value to determine the form behavior
+  const dispatch = useDispatch()
   const challengeType = watch('challengeType')
+
 
   useEffect(() => {
     // Default values for "0v1" and "1v1" challenge types
@@ -52,6 +55,7 @@ const DareChallengeForm = ({
   }
 
   const handleClosePopup = () => {
+    dispatch(setPopupState(false))
     setIsPopupOpen(false)
   }
 
@@ -121,6 +125,7 @@ const DareChallengeForm = ({
     }
     createChallenge(request)
     setIsLoading(true)
+    dispatch(setPopupState(true))
     setIsPopupOpen(true)
   }
